requirements.txt
json5
parse
ply
gitpython
pydantic
dtschema
PyYAML
types-PyYAML
